+++ components/api-server/test/registration-cluster.test.js
@@ -14,6 +14,7 @@ const charlatan = require('charlatan');
 const bluebird = require('bluebird');
 const supertest = require('supertest');
 
+const helpers = require('./helpers');
 const { getConfig } = require('@pryv/boiler');
 const { getApplication } = require('api-server/src/application');
 const ErrorIds = require('errors/src/ErrorIds');
@@ -59,6 +60,7 @@ describe('registration: cluster', function () {
     mongoFixtures = databaseFixture(await produceMongoConnection());
     await mongoFixtures.context.cleanEverything();
   });
+
   before(async function () {
     config.injectTestConfig({
       dnsLess: { isActive: false },
@@ -76,6 +78,7 @@ describe('registration: cluster', function () {
 
     usersRepository = await getUsersRepository();
   });
+
   after(async function () {
     config.injectTestConfig({});
     mongoFixtures = databaseFixture(await produceMongoConnection());
@@ -98,6 +101,7 @@ describe('registration: cluster', function () {
     }
     return validationRequest;
   }
+
   function buildRegistrationRequest (user, request, hasToken = true) {
     const registrationRequest = {
       host: { name: res.req._header.split('Host: ')[1].split('\r\n')[0] },
@@ -115,6 +119,7 @@ describe('registration: cluster', function () {
     }
     return registrationRequest;
   }
+
   function stripRegistrationRequest (request) {
     delete request.user.language;
     delete request.user.id;
@@ -264,6 +269,7 @@ describe('registration: cluster', function () {
         assert.equal(secondEmail, userData.email);
       });
     });
+
     describe('when a user with the same username/email already exists in core but not in register', () => {
       let firstValidationRequest;
       let firstRegistrationRequest;
@@ -331,6 +337,7 @@ describe('registration: cluster', function () {
         assert.deepEqual(registrationSent, firstRegistrationRequest);
       });
     });
+
     describe('when the username exists in register', () => {
       before(async () => {
         userData = _.extend({}, defaults(), { username: 'wactiv' });
@@ -627,6 +634,7 @@ describe('registration: cluster', function () {
         });
       });
     });
+
     describe('when invitationTokens are set to [] (forbidden creation)', () => {
       describe('when any string is provided', () => {
         before(async () => {
@@ -700,6 +708,39 @@ describe('registration: cluster', function () {
             }
           ]);
         });
+      });
+    });
+
+    describe('[RCPW] When password rules are enabled', function () {
+      const validation = helpers.validation;
+
+      before(async () => {
+        config.injectTestConfig(helpers.passwordRules.settingsOverride);
+        userData = defaults();
+        nock(regUrl)
+          .post('/users/validate', () => true)
+          .reply(200, { errors: [] });
+        nock(regUrl)
+          .post('/users', () => true)
+          .reply(200, { username: userData.username });
+        nock(regUrl)
+          .put('/users', () => true)
+          .reply(200, { ok: true });
+      });
+
+      it('[0OBL] must fail if the new password does not comply (smoke test; see "/change-password" in account tests)', async () => {
+        userData.password = helpers.passwordRules.passwords.badTooShort;
+        const res = await request.post(methodPath).send(userData);
+        validation.checkError(res, {
+          status: 400,
+          id: ErrorIds.InvalidParametersFormat
+        });
+      });
+
+      it('[5BQL] must succeed if the new password complies (smoke test; see "/change-password" in account tests)', async () => {
+        userData.password = helpers.passwordRules.passwords.good4CharCats;
+        const res = await request.post(methodPath).send(userData);
+        validation.check(res, { status: 201 });
       });
     });
   });